+++ CHANGELOG.md
@@ -8,6 +8,7 @@
   - **_Breaking Change_**: rewrite `ASTNode`, add `NTNode` and `TNode`.
     - Rewrite selectors and traversers.
   - **_Breaking Change_**: remove `IParserBuilder.lexer`. Set the lexer in parser builders' constructor instead.
+  - **_Breaking Change_**: remove `ASTNode.token`, add `IParserBuilder.mapper` to transform token into `ASTData`. [#38](https://github.com/DiscreteTom/retsac/issues/38)
   - Feat: add `ExtractSerializableParserData`. [#41](https://github.com/DiscreteTom/retsac/issues/41)
   - Feat: add `NTNode.global` to share data across all NT nodes.
     - Add `IParserBuilder.global` to set the global data and cloner.